backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
langchain
langchain-core
langchain-groq
langchain-google-genai
langsmith
groq
Pillow
requests
google-cloud-storage
python-multipart
email-validator
